requirements.txt
Flask
pandas
numpy
joblib
pyserial
# time
# os